Make SampleItems a shared component
